Overhauls UI. Created a StatsWindow GUI using TKinter. Application now opens a window to retrieve the data from the user. Will correctly display the team chosen, and some functionality of the HR leaders button.
Adjusted the return statement of the HR leader function of Teams class. Now has a parameter to choose whether to return a printed statement to be used by the GUI, or a pandas DF which will used at a later data. Default is set to the print statement "df_return=False
"