Changed calls in mlb_plotter to leader_lookup calls instead of specific calls. This allows me to pass the type, instead of calling a different funciton for each stat
Changed hr_leaders function in Teams clas to leader_lookup, adding the stat type parameter to the call. Allows for a single function to be used when calling all the different stats. Consequently added a try/except catch to adjust the data type conversion to fit with what is being returned.